renamed sharedComponents folder to SharedComponents [PT #141915635]
