+++ src/modules/SharedComponents/FileUploader/components/Confirmation.js
