+++ src/modules/SharedComponents/FileUploader/components/Confirmation.scss
