+++ src/modules/SharedComponents/FileUploader/components/Confirmation.snapshot.test.js
