+++ src/modules/SharedComponents/FileUploader/components/FileStepper.scss
