+++ src/modules/SharedComponents/FileUploader/components/FileStepper.snapshot.test.js
