+++ src/modules/SharedComponents/FileUploader/components/FileUpload.scss
