+++ src/modules/SharedComponents/FileUploader/components/FileUploadInfoRow.js
