+++ src/modules/SharedComponents/FileUploader/components/FileUploadInfoRow.scss
