+++ src/modules/SharedComponents/FileUploader/components/FileUploadInfoRow.snapshot.test.js
