+++ src/modules/SharedComponents/FileUploader/components/FileUploader.snapshot.test.js
