+++ src/modules/SharedComponents/FileUploader/components/GettingStarted.js
