+++ src/modules/SharedComponents/FileUploader/components/Metadata.snapshot.test.js
