+++ src/modules/SharedComponents/FileUploader/components/UploadDialog.js
