+++ src/modules/SharedComponents/FileUploader/components/UploadedFileMetadata.js
