+++ src/modules/SharedComponents/FileUploader/components/UploadedFileMetadata.scss
