+++ src/modules/SharedComponents/FileUploader/components/UploadedFileMetadata.snapshot.test.js
