+++ src/modules/SharedComponents/FileUploader/components/fileHelper.js
