+++ src/modules/SharedComponents/FileUploader/containers/Confirmation.js
