+++ src/modules/SharedComponents/FileUploader/containers/UploadedFileMetadata.js
