+++ src/modules/SharedComponents/FileUploader/reducer.js
